// frontend/src/types/model.ts
// Code generated by tygo. DO NOT EDIT.
export interface GormModel {
  ID: number;
  CreatedAt: string;
  UpdatedAt: string;
  DeletedAt: string | null;
}

//////////
// source: user.go

export interface User extends GormModel {
  Name: string;
  BirthDate?: string;
  DeathDate?: string;
  LegalDomicile?: string;
  LastAddress?: string;
  Remarks?: string;
}